+++ ChangeLog.md
@@ -7,6 +7,8 @@ ChangeLog
 * `Parser.tokenize` now returns a Result, giving an Error instead of
 panicking
 * `Parser::new` also returns a Result instead of an Option
+* `detect_leak` now uses `max_distance` to compute a leak, so `leak`
+  parameter is removed.
 
 0.4.0 (2015-06-30)
 ------------------